Resolved all comments raised and added the updateProductFromAdmin to product.controller
